requirements.txt
pycryptodome
pyside2
sm4